MoneyType implemented and tested.
Money consists of an amount and a currency. The currency is implemented trough class Currency(). The Money() class encapsulates Currency() and a Decimal() amount. All useful Money() arithmetics are implemented. Money() can be saved in a database as MoneyField(), which serializes it's contents as a character string. Unit tests for most methods.

TODO: Indexing of MoneyField() has not been taken care of.