Some basic cart and order functionality implemented.
Carts, wishlists, and orders are implemented as one single kind of list containing items. A flag changes the status of a cart into an order (at order time), or adds an entire wishlist to the cart. List items can freeze their price, which happens when their parent list is flagged as order.

In the admin, the lists are split into different views, just like the catalog nodes are split into products and categories (in the catalog app).

There is still a lot of functionality missing, but the basic data model is worth a commit.